Tarea#4/reservas/requirements.txt
flask
pika
pact-python
pytest
requests